Add some better tests and reorganize folder structure a little bit
