+++ package.json
@@ -17,7 +17,8 @@
 		"react-dom": "^16.13.1",
 		"react-redux": "^7.2.0",
 		"react-router-dom": "^5.1.2",
-		"react-scripts": "3.4.1"
+		"react-scripts": "3.4.1",
+		"react-test-renderer": "^16.13.1"
 	},
 	"devDependencies": {
 		"eslint-config-prettier": "^6.0.0",
