+++ src/App.test.js
@@ -3,6 +3,7 @@ import { render } from '@testing-library/react'
 import { Provider } from 'react-redux'
 import store from './app/store'
 import App from './App'
+import renderer from 'react-test-renderer'
 
 test('renders EnterApiKey view by default', () => {
   const { getByText } = render(
@@ -12,4 +13,15 @@ test('renders EnterApiKey view by default', () => {
   )
 
   expect(getByText(/Personal Access Token/i)).toBeInTheDocument()
+})
+
+test('renders App snapshot', () => {
+  const tree = renderer
+    .create(
+      <Provider store={store}>
+        <App />
+      </Provider>
+    )
+    .toJSON()
+  expect(tree).toMatchSnapshot()
 })
