+++ src/ui/components/AppHeader/AppHeader.js
@@ -2,18 +2,23 @@ import React from 'react'
 import { NavLink } from 'react-router-dom'
 import './AppHeader.scss'
 import { destroySession } from '../../../app/store'
+import { useSelector } from 'react-redux'
+import { selectApiKey } from '../../screens/EnterApiKey/state'
 
 const AppHeader = () => {
+  const apiKey = useSelector(selectApiKey)
   return (
     <div className="AppHeader">
       <div className="title">
         <NavLink to="/">Prioritizer</NavLink>
       </div>
-      <div className="logout">
-        <a href="/" onClick={destroySession}>
-          Log out
-        </a>
-      </div>
+      {apiKey && (
+        <div className="logout">
+          <a href="/" onClick={destroySession}>
+            Log out
+          </a>
+        </div>
+      )}
     </div>
   )
 }
