+++ src/ui/screens/LoggedIn/Repo/Issues/state/issuesSlice.js
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit'
-import createClient from '../../../../../../github/client'
+import createClient from '../../../../../../app/githubClient'
 import { updateWithIssues } from './prioritiesSlice'
 
 const initialState = {
